consolidate messages with an SdmxMessage trait
Note: This fixes an error in the original `data_sample01.json` file by aligning it with `data_sample02.json`. Note that `data_sample02.json` accepts a property named "receivers" which is an array.

There is also an error in the original JSON schema for SDMX-JSON Structure Message, where the file `1-sdmx-json-field-guide.md` says that the meta property has a `receiver` type of a Receiver type, even though `structure_sample01.json` and `structure_sample02.json` have a `receivers` property which accept an array of Receiver type.